--- aspnetcore/fundamentals/startup/3.0_samples/Startup2.cs
@@ -1,68 +1,0 @@
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-
-
-namespace WebApplication17
-{
-    #region sample_snapshot
-    public class Startup
-    {
-        private readonly IWebHostEnvironment _env;
-        public IConfiguration Configuration { get; }
-        private readonly ILoggerFactory _loggerFactory;
-
-        public Startup(IWebHostEnvironment env, 
-            IConfiguration configuration,
-            ILoggerFactory loggerFactory)
-        {
-            Configuration = configuration;
-            _env = env;
-            _loggerFactory = loggerFactory;
-        }
-
-        public void ConfigureServices(IServiceCollection services)
-        {
-            var logger = _loggerFactory.CreateLogger<Startup>();
-
-            if (_env.IsDevelopment())
-            {
-                // Development service configuration
-                logger.LogInformation("Development environment");
-            }
-            else
-            {
-                // Non-development service configuration
-                logger.LogInformation($"Environment: {_env.EnvironmentName}");
-            }
-
-            // Configuration is available during startup.
-            // Examples:
-            //   _config["key"]
-            //   _config["subsection:suboption1"]
-        }
-        #endregion
-
-        public void Configure(IApplicationBuilder app,  IWebHostEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-
-            app.UseRouting();
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapGet("/", async context =>
-                {
-                    await context.Response.WriteAsync("Hello World!");
-                });
-            });
-        }
-    }
-}
