+++ aspnetcore/fundamentals/startup/3.0_samples/StartupFilterSample/Pages/Privacy.cshtml
@@ -6,5 +6,5 @@
 <h1>@ViewData["Title"]</h1>
 
 <p>Use this page to detail your site's privacy policy.</p>
-
+<p> Append query string ?option=hello</p>
 Option String: @HttpContext.Items["option"];
