+++ aspnetcore/fundamentals/startup/3.0_samples/StartupFilterSample/Pages/Shared/_Layout.cshtml
@@ -22,7 +22,7 @@
                             <a class="nav-link text-dark" asp-area="" asp-page="/Index">Home</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-page="/Privacy">Privacy</a>
+                            <a class="nav-link text-dark" asp-area="" asp-page="/Privacy" asp-route-option="Hello">Filter</a>
                         </li>
                     </ul>
                 </div>
