+++ aspnetcore/fundamentals/startup/3.0_samples/StartupFilterSample/StartUp2.cs
@@ -1,0 +1,36 @@
+//#define B
+#if B
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+
+namespace StartupFilterSample
+{
+    #region snippet
+    public class Startup
+    {
+        private readonly IWebHostEnvironment _env;
+        public IConfiguration Configuration { get; }
+
+        public Startup(IWebHostEnvironment env, IConfiguration configuration)
+        {
+            Configuration = configuration;
+            _env = env;
+        }
+
+        public void ConfigureServices(IServiceCollection services)
+        {
+            if (_env.IsDevelopment())
+            {
+            }
+            else
+            {
+            }
+
+            //   _config["key"]
+        }
+    }
+    #endregion
+}
+#endif
